Revert "Feat/rss ecomm 3 18 manage addresses (#88)"
This reverts commit 30a4864cdd36239666bc037be3f1be1bfdf9018c.
