--- src/components/addressBlock/address-block.module.scss
@@ -1,9 +1,0 @@
-.addressBlock {
-  @apply flex max-w-[33%] flex-grow flex-col items-center justify-center gap-2 rounded-xl bg-bgContainer p-7;
-}
-.postalCodeBlock {
-  @apply flex w-[100%] justify-between gap-2;
-}
-.defaultLabel {
-  @apply mb-1 flex rounded-lg bg-borderLight p-1 text-textDark;
-}
