--- src/components/changePasswordBlock/changePasswordBlock.module.scss
@@ -1,3 +1,0 @@
-.changePasswordBlock {
-  @apply flex max-w-[50%] flex-grow flex-col items-center justify-center gap-2 rounded-xl bg-bgContainer p-7;
-}
