+++ src/components/input/input-field/input-date-field/input-date-field.ts
@@ -3,7 +3,7 @@ import InputField, { InputFiledProps } from '@components/input/input-field/input
 type InputDateFieldProps = InputFiledProps;
 
 export default class InputDateField extends InputField {
-  constructor({ name, callback, attributes, additionalClassNames }: InputDateFieldProps) {
-    super({ name, callback, attributes, additionalClassNames });
+  constructor({ name, callback, attributes }: InputDateFieldProps) {
+    super({ name, callback, attributes });
   }
 }
