+++ src/pages/profile/profile-page.ts
@@ -6,9 +6,12 @@ import Container from '@components/container/container';
 import CustomerApi from '@api/customerApi';
 import { apiInstance } from '@api/api';
 import { Address, ClientResponse, Customer } from '@commercetools/platform-sdk';
+import InputDateField from '@components/input/input-field/input-date-field/input-date-field';
+import InputTextField from '@components/input/input-field/input-password-field/input-text-field';
+import { validationDate } from '@utils/validation/date';
+import { validationName } from '@utils/validation/name';
 import AddressBlock from '@components/addressBlock/addressBlock';
-import UserInfoBlock from '@components/userInfoBlock/userInfoBlock';
-import ChangePasswordBlock from '@components/changePasswordBlock/changePasswordBlock';
+// import Dropdown from '@components/dropdown/dropdown';
 
 export default class ProfilePage extends View {
   container: HTMLElement;
@@ -16,6 +19,22 @@ export default class ProfilePage extends View {
   customerId: string = localStorage.getItem('customerID') || '';
 
   customerApi: CustomerApi = new CustomerApi(apiInstance);
+
+  dateOfBirthInput: InputDateField;
+
+  lastNameInput: InputTextField;
+
+  firstNameInput: InputTextField;
+
+  isValidName: boolean;
+
+  isValidLastName: boolean;
+
+  isValidDateOfBirth: boolean;
+
+  isValidStreet: boolean;
+
+  isValidCity: boolean;
 
   shippingAddresses: ElementCreator;
 
@@ -26,12 +45,7 @@ export default class ProfilePage extends View {
   shippingTitle: ElementCreator;
 
   billingTitle: ElementCreator;
-  
-  userInfo: ElementCreator;
 
-  addShippingAddress: ElementCreator;
-
-  addBillingAddress: ElementCreator;
   constructor() {
     const params: ParamsElementCreator = {
       tag: 'section',
@@ -39,9 +53,24 @@ export default class ProfilePage extends View {
     };
     super(params);
     this.container = Container.get();
-    this.userInfo = new ElementCreator({
-      tag: 'div',
-      classNames: [styles.userInfo],
+    this.firstNameInput = new InputTextField({
+      name: 'first name',
+      callback: this.validateName.bind(this),
+      attributes: [{ type: 'disabled', value: 'true' }],
+      additionalClassNames: [`${inputStyles.userInputItem}`],
+    });
+    this.lastNameInput = new InputTextField({
+      name: 'last name',
+      callback: this.validateLastName.bind(this),
+      attributes: [{ type: 'disabled', value: 'true' }],
+      additionalClassNames: [`${inputStyles.userInputItem}`],
+    });
+
+    this.dateOfBirthInput = new InputTextField({
+      name: 'Date of birth',
+      callback: this.validateDateOfBirth.bind(this),
+      attributes: [{ type: 'disabled', value: 'true' }],
+      additionalClassNames: [`${inputStyles.userInputItem}`],
     });
     this.shippingAddresses = new ElementCreator({
       tag: 'div',
@@ -50,18 +79,6 @@ export default class ProfilePage extends View {
     this.billingAddresses = new ElementCreator({
       tag: 'div',
       classNames: [`${styles.addressBlock}`],
-    });
-    this.addShippingAddress = new ElementCreator({
-      tag: 'span',
-      textContent: 'Add',
-      callback: [{ event: 'click', callback: () => this.addAddress.call(this, 'shipping') }],
-      classNames: [`${styles.header}`],
-    });
-    this.addBillingAddress = new ElementCreator({
-      tag: 'span',
-      textContent: 'Add',
-      callback: [{ event: 'click', callback: () => this.addAddress.call(this, 'billing') }],
-      classNames: [`${styles.header}`],
     });
     this.shippingTitle = new ElementCreator({
       tag: 'div',
@@ -73,8 +90,6 @@ export default class ProfilePage extends View {
       textContent: 'Billing addresses 💶',
       classNames: [`${styles.header}`],
     });
-    this.shippingTitle.getElement().appendChild(this.addShippingAddress.getElement());
-    this.billingTitle.getElement().appendChild(this.addBillingAddress.getElement());
     this.userAddresses = new ElementCreator({
       tag: 'div',
       classNames: [styles.userAddresses],
@@ -85,18 +100,29 @@ export default class ProfilePage extends View {
         this.billingAddresses.getElement(),
       ],
     });
+
+    this.isValidName = false;
+    this.isValidLastName = false;
+    this.isValidDateOfBirth = false;
+    this.isValidStreet = false;
+    this.isValidCity = false;
     this.getElement().append(this.container);
     this.configureView();
     this.setСustomerInfo();
-    this.getAllAddresses();
   }
 
   private configureView(): void {
     const title = new ElementCreator({ tag: 'div', textContent: 'Your info ⭐', classNames: [`${styles.header}`] });
+    const userInfo = new ElementCreator({
+      tag: 'div',
+      classNames: [styles.userInfo],
+      children: [this.firstNameInput.getElement(), this.lastNameInput.getElement(), this.dateOfBirthInput.getElement()],
+    });
+
     const userInfoBlock = new ElementCreator({
       tag: 'div',
       classNames: [styles.userInfoBlock],
-      children: [this.userInfo.getElement(), this.userAddresses.getElement()],
+      children: [userInfo.getElement(), this.userAddresses.getElement()],
     });
     this.container.append(title.getElement(), userInfoBlock.getElement());
   }
@@ -113,23 +139,10 @@ export default class ProfilePage extends View {
   private async setСustomerInfo(): Promise<void> {
     const customerInfo = await this.getcustomerInfo();
     if (customerInfo) {
-      if (
-        customerInfo.firstName &&
-        customerInfo.lastName &&
-        customerInfo.dateOfBirth &&
-        customerInfo.version &&
-        customerInfo.email
-      ) {
-        const userInfoBlock = new UserInfoBlock({
-          firstName: customerInfo.firstName,
-          lastName: customerInfo.lastName,
-          dateOfBirth: customerInfo.dateOfBirth,
-          email: customerInfo.email,
-        });
-        const changePasswordBlock = new ChangePasswordBlock();
-        this.userInfo.getElement().appendChild(userInfoBlock.getElement());
-        this.userInfo.getElement().appendChild(changePasswordBlock.getElement());
-        localStorage.setItem('customerVersion', `${customerInfo.version}`);
+      if (customerInfo.firstName && customerInfo.lastName && customerInfo.dateOfBirth) {
+        this.firstNameInput.setValue(customerInfo.firstName);
+        this.lastNameInput.setValue(customerInfo.lastName);
+        this.dateOfBirthInput.setValue(customerInfo.dateOfBirth);
       }
       this.setCustomerAddresses({
         addresses: customerInfo.addresses,
@@ -138,27 +151,9 @@ export default class ProfilePage extends View {
         billingAddressIds: customerInfo.billingAddressIds,
         defaultBillingAddressId: customerInfo.defaultBillingAddressId,
       });
-    } else {
-      console.error('Wrong customer data!');
     }
   }
 
-  private addAddress(addressType: string) {
-    const addressParams = {
-      street: '',
-      city: '',
-      postalCode: '',
-      country: 'US',
-      isDefaultShipping: 'no',
-      isDefaultBilling: 'no',
-      addressId: '',
-      isNewAddress: 'yes',
-      addressType: addressType,
-    };
-    const address = new AddressBlock(addressParams).getElement();
-    if (addressType === 'shipping') this.shippingAddresses.getElement().appendChild(address);
-    if (addressType === 'billing') this.billingAddresses.getElement().appendChild(address);
-  }
   private async setCustomerAddresses(
     addressesInfo: Pick<
       Customer,
@@ -179,8 +174,7 @@ export default class ProfilePage extends View {
           currentAddress.streetName &&
           currentAddress.city &&
           currentAddress.postalCode &&
-          currentAddress.country &&
-          currentAddress.id
+          currentAddress.country
         ) {
           const addressParams = {
             street: currentAddress.streetName,
@@ -189,7 +183,6 @@ export default class ProfilePage extends View {
             country: currentAddress.country,
             isDefaultShipping: 'no',
             isDefaultBilling: 'no',
-            addressId: currentAddress.id,
           };
           if (defaultShippingAddressId && currentAddress.id && defaultShippingAddressId.includes(currentAddress.id)) {
             addressParams.isDefaultShipping = 'yes';
@@ -208,7 +201,32 @@ export default class ProfilePage extends View {
     }
   }
 
-  private getAllAddresses() {
-    console.log(this.shippingAddresses.element.childNodes);
+  private validateDateOfBirth(): void {
+    const validationMessages: string[] = validationDate(this.dateOfBirthInput.getValue());
+    this.dateOfBirthInput.setErrors(validationMessages);
+    this.isValidDateOfBirth = !validationMessages.length;
+    this.isAllFieldsValid();
+  }
+
+  private validateName(): void {
+    const validationMessages: string[] = validationName(this.firstNameInput.getValue());
+    this.firstNameInput.setErrors(validationMessages);
+    this.isValidName = !validationMessages.length;
+    this.isAllFieldsValid();
+  }
+
+  private validateLastName(): void {
+    const validationMessages: string[] = validationName(this.lastNameInput.getValue());
+    this.lastNameInput.setErrors(validationMessages);
+    this.isValidLastName = !validationMessages.length;
+    this.isAllFieldsValid();
+  }
+
+  isAllFieldsValid() {
+    if (this.isValidName && this.isValidLastName && this.isValidDateOfBirth) {
+      return true;
+    } else {
+      return false;
+    }
   }
 }
